взаимодействие между процессами

+++ src/DebugServer/DebugeeProcess.cs
@@ -1,11 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.Net.Sockets;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Text;
 using System.Threading.Tasks;
+
+using VSCodeDebug;
 
 namespace DebugServer
 {
@@ -13,8 +16,8 @@ namespace DebugServer
     {
         public DebugeeOutputEventArgs(string category, string content)
         {
-            Category = category;
-            Content = Content;
+            this.Category = category;
+            this.Content = content;
         }
 
         public string Category { get; }
@@ -27,6 +30,9 @@ namespace DebugServer
         private Process _process;
         private TcpClient _client;
 
+        private bool _terminated;
+        private bool _stdoutEOF;
+        private bool _stderrEOF;
 
         public string RuntimeExecutable { get; set; }
         public string WorkingDirectory { get; set; }
@@ -80,23 +86,24 @@ namespace DebugServer
 
         private void Process_Exited(object sender, EventArgs e)
         {
+            Terminate();
             ProcessExited?.Invoke(this, new EventArgs());
-        }
-
-        private void Process_ErrorDataReceived(object sender, DataReceivedEventArgs e)
-        {
-            if (e.Data == null)
-            {
-                //_stderrEOF = true;
-            }
-            SendOutput("stderr", e.Data);
         }
 
         private void Process_OutputDataReceived(object sender, DataReceivedEventArgs e)
         {
             if (e.Data == null)
             {
-                //_stderrEOF = true;
+                _stdoutEOF = true;
+            }
+            SendOutput("stdout", e.Data);
+        }
+
+        private void Process_ErrorDataReceived(object sender, DataReceivedEventArgs e)
+        {
+            if (e.Data == null)
+            {
+                _stderrEOF = true;
             }
             SendOutput("stderr", e.Data);
         }
@@ -106,9 +113,25 @@ namespace DebugServer
             OutputReceived?.Invoke(this, new DebugeeOutputEventArgs(category, data));
         }
 
+        private void Terminate()
+        {
+            if (!_terminated)
+            {
+
+                // wait until we've seen the end of stdout and stderr
+                for (int i = 0; i < 100 && (_stdoutEOF == false || _stderrEOF == false); i++)
+                {
+                    System.Threading.Thread.Sleep(100);
+                }
+                
+                _terminated = true;
+                _process = null;
+            }
+        }
+
         public void Kill()
         {
-            if (!_process.HasExited)
+            if (_process != null && !_process.HasExited)
             {
                 _process.Kill();
             }
